fix: correct v8__String__NewExternalOneByte length parameter type (#1962)
The C++ function declared `int length` (4 bytes) while the Rust FFI
binding declared `size_t` (8 bytes), creating a type mismatch at the
FFI boundary. This aligns it with `v8__String__NewExternalTwoByte`
which already correctly uses `size_t length`.

+++ src/binding.cc
@@ -1275,7 +1275,7 @@ const v8::String* v8__String__NewExternalOneByteStatic(v8::Isolate* isolate,
 }
 
 const v8::String* v8__String__NewExternalOneByte(
-    v8::Isolate* isolate, char* data, int length,
+    v8::Isolate* isolate, char* data, size_t length,
     ExternalOneByteString::RustDestroy rustDestroy) {
   return maybe_local_to_ptr(v8::String::NewExternalOneByte(
       isolate, new ExternalOneByteString(data, length, rustDestroy, isolate)));
